backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-decouple==3.8
email-validator==2.1.0
pymongo==4.5.0
python-dotenv==1.0.0
web3==6.11.1
cryptography==41.0.4
pytest==7.4.3
motor==3.3.1
requests>=2.31.0
eth-account>=0.8.0
eth-utils>=2.1.0 